requirements.txt
urllib
json
python-dotenv